% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BehavioralEconomicR.R
\docType{package}
\name{BehavioralEconomicsR}
\alias{BehavioralEconomicsR}
\alias{BehavioralEconomicsR-package}
\title{Paquete de R disenado para trabajar de forma sencilla data de investigaciones de Economia Conductual. Creado por BEST}
\description{
Paquete de R disenado para trabajar de forma sencilla data de investigaciones de Economia Conductual. Creado por BEST
}
