% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BehavioralEconomicsR.R
\docType{data}
\name{ie_data}
\alias{ie_data}
\title{Paquete de R disenado para trabajar de forma sencilla data de investigaciones de Economia Conductual.}
\format{Base de datos del libro Irrational Exuberance de Roberth Shiller, premio Nobel de Economia por su aportes en finanzas conductuales. Esta data es actualizada mensualmente por el propio autor y la Universidad de Yale.}
\usage{
data(ie_data)
}
\description{
Paquete de R disenado para trabajar de forma sencilla data de investigaciones de Economia Conductual.
}
\keyword{datasets}
